Fix grid texture + add noise to texture

+++ Burning bent world/Assets/Code/Scripts/Utils/Utils.cs
@@ -156,6 +156,6 @@ namespace Utils
         /// <param name="a">The value</param>
         /// <param name="b">The modulus' base</param>
         /// <returns>The result of a mod b</returns>
-        public static int MathModulus(int a, int b) => (Math.Abs(a * b) + a) % b;
+        public static float MathModulus(float a, int b) => (Math.Abs(a * b) + a) % b;
     }
 }
